refactor search logic into search.py, fix api search endpoint input & response
